# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eletro AR - Admin",
        "version": "0.1"
    },
    "paths": {
...
        "/api/service-orders": {
            "get": {
                "tags": [
                    "Service Orders"
                ],
                "summary": "Listar ordens de serviço",
                "description": "Retorna uma lista de ordens de serviço paginadas",
                "operationId": "492f8809f5423cdccb244073160398c4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página a ser recuperada",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "service_orders": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "service_date": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "service_description": {
                                                        "type": "string"
                                                    },
                                                    "client_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "client": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "document": {
                                                                "type": "string"
                                                            },
                                                            "phone": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            },
                                                            "address_id": {
                                                                "type": "integer"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "deleted_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "nullable": true
                                                            },
                                                            "address": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "zip_code": {
                                                                        "type": "string"
                                                                    },
                                                                    "street": {
                                                                        "type": "string"
                                                                    },
                                                                    "number": {
                                                                        "type": "string"
                                                                    },
                                                                    "city": {
                                                                        "type": "string"
                                                                    },
                                                                    "state": {
                                                                        "type": "string"
                                                                    },
                                                                    "complement": {
                                                                        "type": "string"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "date-time"
                                                                    },
                                                                    "updated_at": {
                                                                        "type": "string",
                                                                        "format": "date-time"
                                                                    },
                                                                    "deleted_at": {
                                                                        "type": "string",
                                                                        "format": "date-time",
                                                                        "nullable": true
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Service Orders"
                ],
                "summary": "Criar ordem de serviço",
                "description": "Cria uma nova ordem de serviço",
                "operationId": "f824f2e354fb3ccdad43aca088c92161",
                "requestBody": {
                    "description": "Dados da ordem de serviço",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_date",
                                    "service_description",
                                    "client_id"
                                ],
                                "properties": {
                                    "service_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "service_description": {
                                        "type": "string"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ordem de serviço criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "service_order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/service-orders/{serviceOrder}": {
            "get": {
                "tags": [
                    "Service Orders"
                ],
                "summary": "Detalhes da ordem de serviço",
                "description": "Retorna os detalhes de uma ordem de serviço",
                "operationId": "aca2a4e7b6e305d9a4a44ce7005d9c22",
                "parameters": [
                    {
                        "name": "serviceOrder",
                        "in": "path",
                        "description": "ID da ordem de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "service_order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ordem de serviço não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Service Orders"
                ],
                "summary": "Atualizar ordem de serviço",
                "description": "Atualiza uma ordem de serviço",
                "operationId": "7d85bc1db3924e4dbad99cc132de7a87",
                "parameters": [
                    {
                        "name": "serviceOrder",
                        "in": "path",
                        "description": "ID da ordem de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da ordem de serviço",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_date",
                                    "service_description",
                                    "client_id"
                                ],
                                "properties": {
                                    "service_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "service_description": {
                                        "type": "string"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ordem de serviço atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "service_order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Service Orders"
                ],
                "summary": "Deletar ordem de serviço",
                "description": "Deletar uma ordem de serviço",
                "operationId": "2d932bf4811113ad7ca7b3a1cd93a0c7",
                "parameters": [
                    {
                        "name": "serviceOrder",
                        "in": "path",
                        "description": "ID da ordem de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ordem de serviço deletada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ordem de serviço não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}